requirements.txt
Django==1.8.4
djangorestframework==3.2.4
hawkrest==0.0.8
mohawk==0.3.0
requests==2.7.0
six==1.9.0
virtualenv==13.1.2
wheel==0.24.0
